# Conflicts:
#	src/cmd/ship.go

+++ src/cmd/ship.go
@@ -26,7 +26,7 @@ type shipConfig struct {
 	isShippingInitialBranch bool
 	isOffline               bool
 	branchesDeletedOnRemote []string // local branches whose tracking branches have been deleted
-	pullRequestNumber       int64
+	pullRequestNumber       int
 	deleteOriginBranch      bool
 }
 
@@ -156,7 +156,7 @@ func determineShipConfig(args []string, driver hosting.Driver, repo *git.ProdRep
 	branchToMergeInto := repo.Config.ParentBranch(branchToShip)
 	canShipWithDriver := false
 	defaultCommitMessage := ""
-	pullRequestNumber := int64(-1)
+	pullRequestNumber := -1
 	if hasTrackingBranch && !isOffline && driver != nil {
 		prInfo, err := driver.LoadPullRequestInfo(branchToShip, branchToMergeInto)
 		if err != nil {
